+++ Mapwize/Mapwize/Classes/MWZMapOptions.m
@@ -7,6 +7,7 @@
     self.locationEnabled = YES;
     self.beaconsEnabled = NO;
     self.showUserPositionControl = YES;
+    self.displayFloorControl = YES;
     return self;
 }
 
@@ -39,8 +40,15 @@
     if (_minZoom != nil) {
         [optionsDic setObject:_minZoom forKey:@"minZoom"];
     }
+    if (_displayMarkerOptions != nil) {
+        [optionsDic setObject:[_displayMarkerOptions toDictionary] forKey:@"displayMarkerOptions"];
+    }
+    if (_mainColor != nil) {
+        [optionsDic setObject:_mainColor forKey:@"mainColor"];
+    }
     [optionsDic setObject:@0 forKey:@"useBrowserLocation"];
     [optionsDic setObject:[NSNumber numberWithBool:_showUserPositionControl] forKey:@"showUserPositionControl"];
+    [optionsDic setObject:[NSNumber numberWithBool:_displayFloorControl] forKey:@"displayFloorControl"];
     [optionsDic setObject:@0 forKey:@"zoomControl"];
     
     return optionsDic;