+++ Mapwize/Mapwize/Classes/MWZMapView.h
@@ -73,6 +73,8 @@ __attribute__((deprecated("Use setUserPosition:(MWZUserPosition*) instead")));
 - (void) removePromotedPlace:(MWZPlace*) place;
 - (void) removePromotedPlaceWithId:(NSString*) placeId;
 
+- (void) setExternalPlaces: (NSArray<MWZPlace*>*) externalPlaces;
+
 - (void) addIgnoredPlace:(MWZPlace*) place;
 - (void) addIgnoredPlaceWithId:(NSString*) placeId;
 - (void) setIgnoredPlaces:(NSArray<MWZPlace*>*) places;