+++ Mapwize/Mapwize/Classes/MWZPlace.h
@@ -3,6 +3,8 @@
 #import "MWZStyle.h"
 #import "MWZGeometry.h"
 #import "MWZDirectionPoint.h"
+#import "MWZPlaceType.h"
+#import "MWZVenue.h"
 
 
 @interface MWZPlace : NSObject <MWZDirectionPoint>
@@ -15,6 +17,7 @@
 @property(nonatomic, strong) NSArray<MWZTranslation*>* translations;
 @property(nonatomic, strong) NSNumber* order;
 @property(nonatomic, strong) NSString* placeTypeId;
+@property(nonatomic, strong) MWZPlaceType* placeType;
 @property(nonatomic) BOOL isPublished;
 @property(nonatomic) BOOL isSearchable;
 @property(nonatomic) BOOL isVisible;
@@ -24,8 +27,9 @@
 @property(nonatomic, strong) MWZGeometry* geometry;
 @property(nonatomic, strong) MWZCoordinate* marker;
 @property(nonatomic, strong) MWZCoordinate* entrance;
+@property(nonatomic, strong) MWZVenue* venue;
 
-@property NSDictionary* data;
+@property(nonatomic, strong) NSDictionary* data;
 
 
 - (instancetype)initFromDictionary:(NSDictionary*)dic;
@@ -34,5 +38,6 @@
 
 - (NSDictionary*) toDirectionDictionary;
 - (NSString*) toDirectionStringJSON;
+- (NSDictionary*) toDictionary;
 
 @end