+++ Mapwize/Mapwize/Classes/MWZPlaceType.h
@@ -1,0 +1,14 @@
+#import <Foundation/Foundation.h>
+#import "MWZStyle.h"
+
+@interface MWZPlaceType : NSObject
+
+@property(nonatomic, strong) NSString* identifier;
+@property(nonatomic, strong) NSString* name;
+@property(nonatomic, strong) MWZStyle* style;
+
+- (instancetype)initFromDictionary:(NSDictionary*) dic;
+
+- (NSDictionary*) toDictionary;
+
+@end