+++ Mapwize/Mapwize/Classes/MWZVenue.h
@@ -19,6 +19,8 @@
 
 - (instancetype)initFromDictionary:(NSDictionary*)dic;
 
+- (NSDictionary*) toDictionary;
+
 - (MWZBounds*) getBounds;
 
 @end